Added Universe Class
Added Universe class which will do calculations for bodies, create the animations, and draw the lines (Basically the main of the physics engine). OrbitSim will be the main window and background graphics manager, Body now knows nothing about timelines or animation or anything. Current issue is that no calculations are being made so bodies just vibrate in place